+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 ### Added
 
 - Added compatibility with Craft 5.0.0.
+
+### Removed
+
+- Removed the `expectCT` setting.